Link classes relocated.
